refactor: extract the WithIsScrolled component
